# crates/haneul-rpc-api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Haneul Node Api",
    "description": "REST Api for interacting with the Haneul Blockchain",
    "contact": {
      "name": "Haneul Labs",
      "url": "https://github.com/GeunhwaJeong/haneul"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "unknown"
  },
  "servers": [
    {
      "url": "/v2"
    }
  ],
  "paths": {
...
    "/objects/{object_id}": {
      "get": {
        "tags": [
          "Objects"
        ],
        "description": "[![unstable](https://img.shields.io/badge/api-unstable-red?style=for-the-badge)](#) _Api subject to change; use at your own risk_\n\n",
        "operationId": "GetObject",
        "parameters": [
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/objects/{object_id}/version/{version}": {
      "get": {
        "tags": [
          "Objects"
        ],
        "description": "[![unstable](https://img.shields.io/badge/api-unstable-red?style=for-the-badge)](#) _Api subject to change; use at your own risk_\n\n",
        "operationId": "GetObjectWithVersion",
        "parameters": [
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "simple"
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/objects/{object_id}/dynamic-fields": {
      "get": {
        "tags": [
          "Objects"
        ],
        "description": "[![unstable](https://img.shields.io/badge/api-unstable-red?style=for-the-badge)](#) _Api subject to change; use at your own risk_\n\n",
        "operationId": "ListDynamicFields",
        "parameters": [
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {
              "x-haneul-cursor": {
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/transactions/{transaction}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "description": "[![unstable](https://img.shields.io/badge/api-unstable-red?style=for-the-badge)](#) _Api subject to change; use at your own risk_\n\n",
        "operationId": "GetTransaction",
        "parameters": [
          {
            "in": "path",
            "name": "transaction",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "description": "[![unstable](https://img.shields.io/badge/api-unstable-red?style=for-the-badge)](#) _Api subject to change; use at your own risk_\n\n",
        "operationId": "ListTransactions",
        "responses": {
          "200": {
            "description": "",
            "headers": {
              "x-haneul-cursor": {
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {}
            }
          },
          "410": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Transactions"
        ],
        "description": "[![unstable](https://img.shields.io/badge/api-unstable-red?style=for-the-badge)](#) _Api subject to change; use at your own risk_\n\n",
        "operationId": "ExecuteTransaction",
        "requestBody": {
          "content": {
            "application/bcs": {}
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/transactions/simulate": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "description": "[![unstable](https://img.shields.io/badge/api-unstable-red?style=for-the-badge)](#) _Api subject to change; use at your own risk_\n\n",
        "operationId": "SimulateTransaction",
        "requestBody": {
          "content": {
            "application/bcs": {}
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/transactions/resolve": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "description": "[![unstable](https://img.shields.io/badge/api-unstable-red?style=for-the-badge)](#) _Api subject to change; use at your own risk_\n\n",
        "operationId": "ResolveTransaction",
        "requestBody": {
          "content": {
            "application/json": {}
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
  },
...
}